refactor: 카드 입력 validation 로직 분리
